Put all tests into SimulationTest.java
